// frontend/src/api/gen/schemas/technicalQuestion.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * hackathon-portal-api
 * OpenAPI spec version: 1.0.0
 */

export interface TechnicalQuestion {
  binary: boolean;
  /** @nullable */
  description?: string | null;
  event_id: string;
  id: string;
  max_points: number;
  min_points: number;
  question: string;
}